Move NestedDict to a more general location.
